Choose the best model

bestmodel will find the best models from a quickmodel object.

bestmodel(quickmodel)

Arguments

quickmodel

A quickmodel object.

Value

The best model based on the evaluation metric.

Details

This function will take the result from quickmodel and pick the best model. It will pick by the lowest RMSE for regression and highest Accuracy for classification. Notes: This function now only considers RMSE and Accuracy as the metric.

Examples

# classification model
data(PIMA)
class <- quickmodel(Diabetes~., data = PIMA)
#> Loading required package: caret
#> Loading required package: ggplot2
#> Loading required package: lattice
#> Loading required package: gbm
#> Loaded gbm 2.1.8.1
bm <- bestmodel(class)
#> The best model based on Accuracy is gbm
#> 
#>    Model  Accuracy
#> 1 glmnet 0.7697649
#> 2    gbm 0.7722936
#> 3  rpart 0.7463619
#> 4     rf 0.7693161
#> 5    glm 0.7704371

# regression model
data(Boston)
reg <- quickmodel(medv~., data = Boston)
#> Warning: There were missing values in resampled performance measures.
bm <- bestmodel(reg)
#> The best model based on RMSE is rf
#> 
#>    Model     RMSE
#> 1 glmnet 5.209091
#> 2    gbm 3.846278
#> 3  rpart 7.059504
#> 4     rf 3.692132
#> 5     lm 5.034929