Build multiple models

quickmodel will build multiple machine learning models with one function call.

quickmodel(
  formula,
  data,
  metric,
  methods,
  trControl = trainControl(),
  tuneGrid = NULL,
  tuneLength = 3,
  partition = 0.8,
  seed = 1234,
  ...
)

Arguments

formula

An object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted.

data

Data frame from which variables specified in formula are preferentially to be taken.

metric

A string that specifies what summary metric will be used to select the optimal model. By default, possible values are "RMSE" and "Rsquared" for regression and "Accuracy" for classification.

methods

A character vector specifying which classification or regression model to use. By default, possible values are c("lm", "knn", "rf", "rpart", "gbm", "glmnet") for regression and c("glm", "knn", "rf", "rpart", "gbm", "glmnet") for classification.

trControl

A list of values that define how this function acts. Default: trainControl()

tuneGrid

A data frame with possible tuning values. The columns are named the same as the tuning parameters in each method preceded by a period (e.g. .decay, .lambda)., Default: NULL

tuneLength

an integer denoting the number of levels for each tuning parameter that should be generated by createGrid. (NOTE: If given, this argument must be named.), Default: 3

partition

The partition rate. Default: 0.8

seed

The seed to do a random partition. Default: 1234

...

Optional parameters for train in caret

Value

A list of class quickmodel with 5 components:

train

The train dataset

test

The test dataset

models

A list of models

methods

A character vector of machine learning models

metric

The selected metric. Default: "RMSE" for regression and "Accuracy" for classification.

Examples

data("PIMA", package="regclass")
data(Boston, package="MASS")